Do not fail on empty directory when cleaning and test cleaning

+++ src/main/java/de/dagere/peass/ci/clean/callables/CleanRCACallable.java
@@ -6,12 +6,22 @@ import java.io.IOException;
 import org.apache.commons.io.FileUtils;
 import org.jenkinsci.remoting.RoleChecker;
 
+import de.dagere.peass.ci.process.JenkinsLogRedirector;
 import de.dagere.peass.folders.CauseSearchFolders;
 import de.dagere.peass.folders.ResultsFolders;
 import hudson.FilePath.FileCallable;
+import hudson.model.TaskListener;
 import hudson.remoting.VirtualChannel;
 
 public class CleanRCACallable implements FileCallable<Boolean> {
+
+   private static final long serialVersionUID = 2008970638274618905L;
+
+   private final TaskListener listener;
+
+   public CleanRCACallable(final TaskListener listener) {
+      this.listener = listener;
+   }
 
    @Override
    public void checkRoles(final RoleChecker checker) throws SecurityException {
@@ -20,13 +30,15 @@ public class CleanRCACallable implements FileCallable<Boolean> {
 
    @Override
    public Boolean invoke(final File potentialSlaveWorkspace, final VirtualChannel channel) throws IOException, InterruptedException {
-      try {
+      try (final JenkinsLogRedirector redirector = new JenkinsLogRedirector(listener)) {
          String projectName = potentialSlaveWorkspace.getName();
          File folder = new File(potentialSlaveWorkspace.getParentFile(), projectName + "_fullPeass");
          cleanFolder(projectName, folder);
 
          return true;
       } catch (IOException e) {
+         listener.getLogger().println("Exception thrown");
+         e.printStackTrace(listener.getLogger());
          e.printStackTrace();
          return false;
       }
@@ -43,13 +55,18 @@ public class CleanRCACallable implements FileCallable<Boolean> {
       FileUtils.deleteDirectory(resultsFolders.getRCALogFolder());
 
       CauseSearchFolders folders = resultsFolders.getPeassFolders();
-      if (folders.getRcaFolder().exists()) {
-         System.out.println("Deleting: " + folders.getRcaFolder());
-         FileUtils.cleanDirectory(folders.getRcaFolder());
+      if (folders != null) {
+         if (folders.getRcaFolder().exists()) {
+            System.out.println("Deleting: " + folders.getRcaFolder());
+            FileUtils.cleanDirectory(folders.getRcaFolder());
+         }
+
+         System.out.println("Deleting: " + folders.getRCALogFolder());
+         FileUtils.cleanDirectory(folders.getRCALogFolder());
+      } else {
+         System.err.println("Project folder was not existing - not cleaning");
       }
 
-      System.out.println("Deleting: " + folders.getRCALogFolder());
-      FileUtils.cleanDirectory(folders.getRCALogFolder());
    }
 
 }
