+++ src/main/java/de/dagere/peass/ci/clean/callables/CleanUtil.java
@@ -9,11 +9,17 @@ import de.dagere.peass.folders.PeassFolders;
 
 public class CleanUtil {
    public static void cleanProjectFolder(final File folder, final String projectName) throws IOException {
-      PeassFolders folders = new PeassFolders(new File(folder, projectName));
-      FileUtils.deleteDirectory(folders.getProjectFolder());
-      if (folders.getPeassFolder().exists()) {
-         System.out.println("Deleting " + folders.getPeassFolder().getAbsolutePath());
-         FileUtils.deleteDirectory(folders.getPeassFolder());
+      File projectFolder = new File(folder, projectName);
+      if (projectFolder.exists()) {
+         PeassFolders folders = new PeassFolders(projectFolder);
+         FileUtils.deleteDirectory(folders.getProjectFolder());
+         if (folders.getPeassFolder().exists()) {
+            System.out.println("Deleting " + folders.getPeassFolder().getAbsolutePath());
+            FileUtils.deleteDirectory(folders.getPeassFolder());
+         }
+      } else {
+         System.err.println("Project folder " + projectFolder.getAbsolutePath() + " did not exist; did not clean it");
       }
+
    }
 }
